Move asset management, cleanup other files
