+++ projs/shadow/shadow-engine/assets/inc/shadow/assets/management/delegate.h
