+++ projs/shadow/shadow-engine/assets/inc/shadow/assets/management/synchronization.h
