+++ projs/shadow/shadow-engine/assets/inc/shadow/assets/str/string.h
