--- projs/shadow/shadow-engine/assets/src/SFFParser.h
@@ -1,22 +1,0 @@
-#pragma once
-
-#include <string>
-#include <iostream>
-
-#include "SFFElement.h"
-#include "SFFVersion.h"
-
-namespace Shadow::SFF {
-
-	class SFFParser
-	{
-	public:
-
-		static SFFElement* ReadFromStream(std::istream& stream);
-
-		static SFFVersion ReadVersionFromHeader(std::istream& stream);
-
-		static SFFElement* ReadFromFile(std::string path);
-	};
-
-}
