+++ projs/shadow/shadow-engine/assets/src/fs/iostream.cpp
@@ -1,5 +1,6 @@
-#include <fs/iostream.h>
+
 #include <cstring>
+#include "shadow/assets/fs/iostream.h"
 
 namespace ShadowEngine {
