+++ projs/shadow/shadow-engine/assets/src/str/string.cpp
@@ -1,4 +1,6 @@
-#include <str/string.h>
+#include <string>
+
+
 
 namespace ShadowEngine::Str {
     std::string toLower(const std::string& str) {
