+++ projs/shadow/shadow-engine/core/inc/shadow/core/ShadowWindow.h
@@ -6,16 +6,12 @@ namespace SH {
 
   class ShadowWindow {
   public:
-
     int Height;
     int Width;
 
     SDL_Window *sdlWindowPtr;
 
     SDL_Surface *sdlSurface = NULL;
-
-
-    //ShadowEngine::Ref<ShadowEngine::Rendering::GraphicsContext> context;
 
     ShadowWindow(int W, int H);
